fix: camera now rotates about fixed point
The camera is now given a child which it will rotate about. This allows
for rotation directly based on what the camera is looking at. However
I'm not entirely convinced that this is the right appproach. Perhaps the
fixed points position should not be locked with the camera's position.
This aligns with how Unity Editor behaves as the editor's camera rotates
about the exact same point regardless of zoom level and only changes
once the camera moves position. Not sure how to implement that however.