feat: clickable objects can transition camera view
If you want an object to have a custom view point for the camera to be
at, you can create an empty GameObject as a child of your object such
that it has the exact rotation and position that you want the camera to
be in. Then pass all objects into the SerializeField under PointManager
and upon click of any of these objects, the camera will be spherically
interpolated towards that object's viewpoint's position and rotation.